Improve shells configurations and scripts.
Remove unnecessary comments in .bash_profile and fix the order of the
files it loads and rename it to .profile.
Add sourcing of .config/user-dirs.dirs in .zshenv and use it in
.bin/capscr and .bin/screencast.
Make .bin/setup print git related info only if it's a git repository.
Add settings for less.
Make EDITOR and VISUAL variables be "nvim" only if NeoVim exists.
Launch ssh-agent in .zshrc only if it doesn't running already on
startup.
Remove insync function from .zlogin.
Remove start-wifi script in favor of a better /etc/network/interfaces
and usage of dhcpcd in arch-linux.