Use a file-per-program layout for aliases & functions
Use dedicated binaries for some functions - path / urls / vinfo which
may become useful in scripting.
Use an `if ! _command_exists` and `return` to exit those scripts when
the relevant commands just don't exist in the system and therefor the
shell functions defined later are not needed.
Update .profile to include these as well.